requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==22.3.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
discord-ext-ui==3.1.9
discord-handler==0.0.5
discord.py @ git+https://github.com/Rapptz/discord.py.git@30c7a2179f45d370d4d2591c54943267d4fd1078
dispander @ git+https://github.com/sushi-chaaaan/dispander.git@cde35b1ebf9b31c0ebdb03ceafdc32187a6bf714
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
orjson==3.6.8
pathspec==0.9.0
pip-autoremove==0.10.0
pipdeptree==2.2.1
platformdirs==2.5.2
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.27.1
SongDBCore @ git+https://github.com/sushi-chaaaan/song-db-client@02d58f117d35744c20458c3c679f81d59b87423e
tomli==2.0.1
urllib3==1.26.9
yarl==1.7.2
